requirements.txt
# 核心依赖（语音转文本需要）
openai-whisper>=20231117

# 照片 EXIF 提取
Pillow>=10.0.0
